Lastest Tweaks
Added control of an old animatronic pirate. Added random time for fog
and pirate to talk. Activated around every 15 minutes.  It’s a bit
kludgy and I need to fix it. But it was a last minute hack.

Side note:  After these tweaks I had no trick-or-treaters this year!
But it was fun setting it up and watching it trigger. Nearly got ground
fog to work.